REORGANISATION DATESET/DATASOURCE/VISUALISATION

+++ ai/src/models/health_fact.py
@@ -2,9 +2,9 @@ from database import db
 
 class HealthFact(db.Model):
     __tablename__ = "health_facts"
-    id = db.Column(db.Integer, primary_key=True)
-    indicator_id = db.Column(db.Integer, db.ForeignKey('indicators.id'))
-    location_id = db.Column(db.Integer, db.ForeignKey('locations.id'))
+    id = db.Column(db.BigInteger, primary_key=True)
+    indicator_id = db.Column(db.BigInteger, db.ForeignKey('indicators.id'))
+    location_id = db.Column(db.BigInteger, db.ForeignKey('locations.id'))
     period = db.Column(db.Date)
     value = db.Column(db.Float)
     source = db.Column(db.String)
